+++ virtual_machine/instruction_sub.c
@@ -14,7 +14,7 @@
 
 int 	sub(t_struct *data, t_pc *p)
 {
-	unsigned int 	reg[3];
+	long int 		reg[3];
 	unsigned char 	*args;
 	unsigned char 	*args_len;
 	unsigned char	*point;
